# static/openapi.yaml
openapi: 3.0.3
info:
  title: Api-SportMatch
  description: >
    API REST de SportMatch. Fase 2: autenticación, usuarios y administradores.
    Fase 3 (en construcción): deportes, cuestionario de registro, eventos,
    inscripciones y valoraciones, organizadores y validación, administración
    y soporte. Ver /docs/entidades_completas.md en el repositorio para el
    catálogo completo de las 89 entidades del modelo.
  version: 0.5.0
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/rutas:
    get:
      summary: Listar rutas públicas. Filtros - deporte_id, creador_id
      tags: [Rutas]
      security: []
      responses:
        "200": { description: Lista de rutas }
    post:
      summary: Crear una ruta propia, opcionalmente con sus puntos GPS (cualquier usuario autenticado)
      tags: [Rutas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, deporte_id]
              properties:
                nombre: { type: string }
                deporte_id: { type: integer }
                distancia_m: { type: number }
                puntos:
                  type: array
                  items:
                    type: object
                    properties:
                      tipo: { type: string, enum: [inicio, fin, control, paso] }
                      latitud: { type: number }
                      longitud: { type: number }
      responses:
        "201": { description: Ruta creada }
  /api/rutas/{ruta_id}:
    get:
      summary: Detalle de una ruta con sus puntos
      tags: [Rutas]
      security: []
      parameters:
        - { name: ruta_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Ruta }
    put:
      summary: Editar mi ruta (dueño o admin)
      tags: [Rutas]
      parameters:
        - { name: ruta_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Ruta actualizada }
        "403": { description: No tienes permisos sobre esta ruta }
    delete:
      summary: Eliminar mi ruta (borrado suave; dueño o admin)
      tags: [Rutas]
      parameters:
        - { name: ruta_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Ruta eliminada }
  /api/rutas/categorias/{categoria_id}/rutas:
    get:
      summary: Ver las rutas ligadas a una categoría de evento
      tags: [Rutas]
      security: []
      parameters:
        - { name: categoria_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Lista de rutas ligadas }
    post:
      summary: Ligar una ruta existente a una categoría de evento (solo admin)
      tags: [Rutas]
      parameters:
        - { name: categoria_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ruta_id]
              properties:
                ruta_id: { type: integer }
      responses:
        "201": { description: Ruta ligada a la categoría }
...